+++ changelog.md
@@ -1,3 +1,6 @@
 # Changelog
 
-Snapshot 2021-09-18 10:49
+**./script/options.gs**
+* Declared 'textFonts' object - Sets font for different elements.
+	* Uses same elements as 'textColours', including a global font.
+* Wrote GET function to expose 'textFonts' (getTextFonts)